Split tower_preview
